Merge catalog-ui, instances-ui, brokers-ui deployments

+++ scripts/prepare-react-components.sh
@@ -32,5 +32,15 @@ cd $CWD
 # copy lib index.js for DOCKERFILE
 mkdir $TEMP_FOLDER
 cp ../${REACT_COMPONENTS_FOLDER}/lib/index.js ./${TEMP_FOLDER}/index.js
-rm -rf ./node_modules/@kyma-project/react-components/lib/index.js
-cp ./${TEMP_FOLDER}/index.js ./node_modules/@kyma-project/react-components/lib/index.js
+
+if [ -d "./node_modules/@kyma-project/react-components/lib/" ]; then
+    rm -rf ./node_modules/@kyma-project/react-components/lib/index.js
+    cp ./${TEMP_FOLDER}/index.js ./node_modules/@kyma-project/react-components/lib/index.js
+fi
+
+for d in */; do
+    if [ -d "./"$d"node_modules/@kyma-project/react-components/lib/" ]; then
+        rm -rf "./"$d"node_modules/@kyma-project/react-components/lib/index.js"
+        cp ./${TEMP_FOLDER}/index.js "./"$d"node_modules/@kyma-project/react-components/lib/index.js"
+    fi 
+done
